New transformation 'ConstSymbolToMap'
Add transformation to extract inter-state symbols used as constants
in other states. This pattern is recurrent in sparse-matrix vector
multiplication, in which an array stores the offset of row start.
By moving the definition of these symbols into a map, it is possible
to expand the nested sdfg and obtain a flat sdfg representation.