Update test case
Test case for ConstSymbolToMap transformation was updated
to limit the size of transient array (temporary storage)
based on the '__start,__stop' symbols defined in map scope.
With one additional assertion on the list of arguments,
the test fails because simplify is propagating map symbols
to the argument list, which should not happen.